Add device option: Set new device to DHCP before editing it (#12)
This means that as soon as you select a device for configuration, its
config is already saved, so if you just want DHCP, you can then bounce
out of the config menu and your device config will still be saved. No
need to switch between static and DHCP just to force it to save its
config.